backend/requirements.txt
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dateutils==0.6.7
Django==2.2.7
django-environ==0.4.5
django-cors-middleware==1.3.1
django-rest-framework-braces==0.3.4
django-rest-swagger==2.2.0
djangorestframework==3.9.2
drf-yasg==1.15.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==19.2
psycopg2-binary==2.8.4
pyparsing==2.4.2
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
